Unify test helpers module name to “test”
There was a mix of testing and test. Follow Django convention:
```python
from django.test import …
```
