+++ itou/openid_connect/inclusion_connect/test.py
@@ -1,7 +1,6 @@
 from django.test import override_settings
 
-from itou.utils.test import TestCase
-from itou.utils.testing import reload_module
+from itou.utils.test import TestCase, reload_module
 
 from . import constants
