+++ itou/utils/test.py
@@ -1,5 +1,10 @@
+import importlib
+import io
+
+import openpyxl
 from bs4 import BeautifulSoup
 from django.test import Client, TestCase as BaseTestCase
+from django.test.utils import TestContextDecorator
 
 
 def format_html(response, **selectors):
@@ -53,3 +58,26 @@ class NoInlineClient(Client):
 
 class TestCase(BaseTestCase):
     client_class = NoInlineClient
+
+
+class reload_module(TestContextDecorator):
+    def __init__(self, module):
+        self._module = module
+        self._original_values = {key: getattr(module, key) for key in dir(module) if not key.startswith("__")}
+        super().__init__()
+
+    def enable(self):
+        importlib.reload(self._module)
+
+    def disable(self):
+        for key, value in self._original_values.items():
+            setattr(self._module, key, value)
+
+
+def get_rows_from_streaming_response(response):
+    """Helper to read streamed XLSX files in tests"""
+
+    content = b"".join(response.streaming_content)
+    workbook = openpyxl.load_workbook(io.BytesIO(content))
+    worksheet = workbook.active
+    return [[cell.value or "" for cell in row] for row in worksheet.rows]
