+++ itou/utils/test_reload_module.py
@@ -3,9 +3,7 @@ import sys
 from django.conf import settings
 from django.test import override_settings
 
-from itou.utils.test import TestCase
-
-from .testing import reload_module
+from itou.utils.test import TestCase, reload_module
 
 
 SOME_VALUE = settings.SECRET_KEY  # is mandatory
