--- itou/utils/testing.py
@@ -1,28 +1,0 @@
-import importlib
-import io
-
-import openpyxl
-from django.test.utils import TestContextDecorator
-
-
-class reload_module(TestContextDecorator):
-    def __init__(self, module):
-        self._module = module
-        self._original_values = {key: getattr(module, key) for key in dir(module) if not key.startswith("__")}
-        super().__init__()
-
-    def enable(self):
-        importlib.reload(self._module)
-
-    def disable(self):
-        for key, value in self._original_values.items():
-            setattr(self._module, key, value)
-
-
-def get_rows_from_streaming_response(response):
-    """Helper to read streamed XLSX files in tests"""
-
-    content = b"".join(response.streaming_content)
-    workbook = openpyxl.load_workbook(io.BytesIO(content))
-    worksheet = workbook.active
-    return [[cell.value or "" for cell in row] for row in worksheet.rows]
